Refactor: 단일영상 컴포넌트
